Create a separate file for Menus classes
